提交代码

+++ user-service/src/main/java/com/cgs/utils/UserIdGenerateUtils.java
@@ -8,11 +8,11 @@ public class UserIdGenerateUtils {
 
     private static AtomicInteger IntId = new AtomicInteger(0);
 
-    public static int getUserId(){
+    public static long getUserId(){
         return (int)(System.currentTimeMillis()/1000L & 0xFFFFFFFF << 16 | id.addAndGet(1) & 0xFFFF);
     }
 
-    public static int getInitId(){
-        return ((int)(System.currentTimeMillis() - 1000000000000L)/1000)<<16L | (IntId.addAndGet(1));
+    public static long getInitId(){
+        return ((System.currentTimeMillis() - 1000000000000L)/1000)<<16L | (IntId.addAndGet(1));
     }
 }
